Refactoring classes into packages
