+++ app/src/main/java/com/alphelios/superrich/MainActivity.kt
@@ -4,9 +4,11 @@ import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import com.alphelios.iap.BillingEventListener
-import com.alphelios.iap.DataWrappers.*
-import com.alphelios.iap.DataWrappers.ErrorType.*
 import com.alphelios.iap.IapConnector
+import com.alphelios.iap.model.BillingResponse
+import com.alphelios.iap.model.PurchaseInfo
+import com.alphelios.iap.model.SkuInfo
+import com.alphelios.iap.type.ErrorType.*
 import com.alphelios.superrich.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity(R.layout.activity_main) {
@@ -75,7 +77,7 @@ class MainActivity : AppCompatActivity(R.layout.activity_main) {
                     ITEM_ALREADY_OWNED_ERROR -> TODO()
                     CONSUME_ERROR -> TODO()
                     ACKNOWLEDGE_ERROR -> TODO()
-                    PURCHASED_PRODUCT_FETCH_ERROR -> TODO()
+                    FETCH_PURCHASED_PRODUCTS_ERROR -> TODO()
                     BILLING_ERROR -> TODO()
                 }
             }
