+++ iap/src/main/java/com/alphelios/iap/model/PurchaseInfo.kt
@@ -1,0 +1,11 @@
+package com.alphelios.iap.model
+
+import com.android.billingclient.api.Purchase
+
+data class PurchaseInfo(
+    val skuInfo: SkuInfo,
+    val purchase: Purchase
+) {
+    val skuId = skuInfo.skuId
+    val skuProductType = skuInfo.skuProductType
+}
