+++ iap/src/main/java/com/alphelios/iap/type/ErrorType.kt
@@ -1,0 +1,6 @@
+package com.alphelios.iap.type
+
+enum class ErrorType {
+    CLIENT_NOT_READY_ERROR, CLIENT_DISCONNECTED, ITEM_ALREADY_OWNED_ERROR, CONSUME_ERROR, ACKNOWLEDGE_ERROR, FETCH_PURCHASED_PRODUCTS_ERROR,
+    BILLING_ERROR
+}
