+++ iap/src/main/java/com/alphelios/iap/type/SkuProductType.kt
@@ -1,0 +1,5 @@
+package com.alphelios.iap.type
+
+enum class SkuProductType {
+    CONSUMABLE, NON_CONSUMABLE, SUBSCRIPTION
+}
